fixed: seperated draggable component
