+++ client/components/Draggable.js
@@ -1,0 +1,79 @@
+import React, {Component} from 'react';
+import { StyleSheet, Text, View, PanResponder, Animated, Image } from 'react-native';
+import { StackNavigator } from 'react-navigation';
+import target from '../../public/img/target.png';
+import SocketIOClient from 'socket.io-client';
+
+export default class Draggable extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      showDraggable: true,
+      dropAreaValues: null,
+      pan: new Animated.ValueXY(),
+      opacity: new Animated.Value(1)
+    };
+    this.socket = SocketIOClient('http://localhost:3000')
+    this.socket.on('connection', () => console.log('connected'))
+  }
+
+  componentWillMount() {
+    this._val = { x:0, y:0 }
+    this.state.pan.addListener((value) => this._val = value);
+
+    this.panResponder = PanResponder.create({
+        onStartShouldSetPanResponder: (e, gesture) => true,
+        onPanResponderGrant: (e, gesture) => {
+          this.state.pan.setOffset({
+            x: this._val.x,
+            y:this._val.y
+          })
+          this.state.pan.setValue({ x:0, y:0})
+        },
+        onPanResponderMove: Animated.event([ 
+          null, { dx: this.state.pan.x, dy: this.state.pan.y }
+        ]),
+        onPanResponderRelease: (e, gesture) => {
+          if (this.isDropArea(gesture)) {
+            console.log("ITEM IS DROPPED")
+            Animated.timing(this.state.opacity, {
+              toValue: 0,
+              duration: 1000
+            }).start(() =>
+              this.setState({
+                showDraggable: false
+              })
+            );
+          } 
+        }
+      });
+  }
+
+  isDropArea(gesture) {
+    return gesture.moveY < 200;
+  }
+
+  render() {
+    return (
+      <View style={{ width: "20%", alignItems: "center" }}>
+        {this.renderDraggable()}
+      </View>
+    );
+  }
+
+  renderDraggable() {
+    const panStyle = {
+      transform: this.state.pan.getTranslateTransform()
+    }
+    if (this.state.showDraggable) {
+      return (
+        <View style={{ position: "absolute" }}>
+          <Animated.View {...this.panResponder.panHandlers} style={[panStyle, {opacity:this.state.opacity}]}>
+            <Image source={target}/>
+          </Animated.View>
+        </View>
+      );
+    }
+  }
+}
