+++ package.json
@@ -32,6 +32,7 @@
     "react": "16.2.0",
     "react-native": "0.52.0",
     "react-native-draggable": "^1.0.12",
+    "react-native-orientation": "^3.1.3",
     "react-navigation": "^1.0.0-beta.27",
     "redux-socket.io-middleware": "^1.0.4",
     "sequelize": "^4.32.2",
